Added Stripe for payments with cards Payments with Stripe complies with EU regulation and PCI DSS regulations so this site would now be able to take real money from customers if it sold real items minor QOL fixes

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { IBasket } from '../../shared/models/basket';
 import { IOrderToCreate } from '../../shared/models/order';
@@ -8,6 +8,7 @@ import { BasketService } from '../../basket/basket.service';
 import { CheckoutService } from '../checkout.service';
 import { ToastrService } from 'ngx-toastr';
 import { NavigationExtras, Router } from '@angular/router';
+import { loadStripe, Stripe, StripeCardCvcElement, StripeCardExpiryElement, StripeCardNumberElement } from '@stripe/stripe-js'
 
 @Component({
   selector: 'app-checkout-payment',
@@ -16,16 +17,61 @@ import { NavigationExtras, Router } from '@angular/router';
 })
 export class CheckoutPaymentComponent implements OnInit{
   @Input() checkoutForm?: FormGroup;
+  @ViewChild('cardNumber') cardNumberElement?: ElementRef;
+  @ViewChild('cardExpiry') cardExpiryElement?: ElementRef;
+  @ViewChild('cardCvc') cardCvcElement?: ElementRef;
+  stripe: Stripe | null = null;
+  cardNumber?: StripeCardNumberElement;
+  cardExpiry?: StripeCardExpiryElement;
+  cardCvc?: StripeCardCvcElement;
+  cardNumberComplete = false;
+  cardExpiryComplete = false;
+  cardCvcComplete = false;
+  cardErrors: any;
   loading = false;
   
   constructor(private basketService: BasketService, private checkoutService: CheckoutService,
     private toastr: ToastrService, private router: Router) {}
 
   ngOnInit(): void {
+    loadStripe('pk_test_51QSJjhGA0lPjhM6ToVDnhiobdTproewEH0m2MYSkFrIOyaTZx0ow6CLPou1Ox4s3FMByhxYugo8doiO0a095KJYp00mLWHpPAZ')
+      .then(stripe => {
+        this.stripe = stripe;
+        const elements = stripe?.elements();
+        if (elements) {
+          this.cardNumber = elements.create('cardNumber');
+          this.cardNumber.mount(this.cardNumberElement?.nativeElement);
+          this.cardNumber.on('change', event => {
+            this.cardNumberComplete = event.complete;
+            if (event.error) this.cardErrors = event.error.message;
+            else this.cardErrors = null
+          })
+
+          this.cardExpiry = elements.create('cardExpiry');
+          this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
+          this.cardExpiry.on('change', event => {
+            this.cardExpiryComplete = event.complete;
+            if (event.error) this.cardErrors = event.error.message;
+            else this.cardErrors = null
+          })
+
+          this.cardCvc = elements.create('cardCvc');
+          this.cardCvc.mount(this.cardCvcElement?.nativeElement);
+          this.cardCvc.on('change', event => {
+            this.cardCvcComplete = event.complete;
+            if (event.error) this.cardErrors = event.error.message;
+            else this.cardErrors = null
+          })
+        }
+      }
+    );
   }
 
   get paymentFormComplete() {
-    return this.checkoutForm?.get('paymentForm')?.valid;
+    return this.checkoutForm?.get('paymentForm')?.valid
+      && this.cardNumberComplete
+      && this.cardExpiryComplete
+      && this.cardCvcComplete;
   }
 
   async submitOrder() {
@@ -34,15 +80,35 @@ export class CheckoutPaymentComponent implements OnInit{
     if (!basket) throw new Error('Cannot Get Basket');
     try {
       const createdOrder = await this.createOrder(basket);
-      this.basketService.deleteBasket(basket);
-      const navigationExtras: NavigationExtras = { state: createdOrder };
-      this.router.navigate(['checkout/success'], navigationExtras);
+      const paymentResult = await this.confirmPaymentWithStripe(basket);
+      console.log(paymentResult);
+      if (paymentResult.paymentIntent) {
+        this.basketService.deleteBasket(basket);
+        const navigationExtras: NavigationExtras = { state: createdOrder };
+        this.router.navigate(['checkout/success'], navigationExtras);
+      } else {
+        this.toastr.error(paymentResult.error.message);
+      }
     } catch (error: any) {
       console.log(error);
       this.toastr.error(error.message);
     } finally {
       this.loading = false;
     }
+  }
+
+  private async confirmPaymentWithStripe(basket: IBasket | null) {
+    if (!basket) throw new Error('Basket is null');
+    const result = this.stripe?.confirmCardPayment(basket.clientSecret!, {
+      payment_method: {
+        card: this.cardNumber!,
+        billing_details: {
+          name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+        }
+      }
+    });
+    if (!result) throw new Error('Problem attempting payment with stripe');
+    return result;
   }
 
   private async createOrder(basket: IBasket | null) {
@@ -61,5 +127,4 @@ export class CheckoutPaymentComponent implements OnInit{
       shipToAddress: shipToAddress
     }
   }
-
 }
